feat: update Dashboard page
