+++ sql/create_table.sql
@@ -20,7 +20,7 @@ CREATE TABLE IF NOT EXISTS tbl_movieinfo (
 -- 上映館テーブル
 CREATE TABLE IF NOT EXISTS tbl_theater (
   id INT UNSIGNED NOT NULL UNIQUE AUTO_INCREMENT COMMENT 'ID',
-  theater VARCHAR(246) NOT NULL UNIQUE COMMENT '上映館',
+  name VARCHAR(246) NOT NULL UNIQUE COMMENT '上映館',
 
   PRIMARY KEY (id)
 );
